requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.5
django-crispy-forms==1.12.0
django-filter==2.4.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
gensim==4.0.1
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
joblib==1.0.1
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.0.1
nltk==3.6.2
numpy==1.21.1
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.3.2
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.8.2
python-Levenshtein==0.12.2
python-magic==0.4.24
python-magic-bin==0.4.14
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
scipy==1.7.1
six==1.16.0
smart-open==5.1.0
sqlparse==0.4.1
stop-words==2018.7.23
tablib==3.0.0
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.6
xlrd==2.0.1
xlwt==1.3.0
djangorestframework-jwt
django-cors-headers